Move files to new folder structure
